+++ src/redturtle/prenotazioni/adapters/stringinterp.py
@@ -23,8 +23,8 @@ except ImportError:
 @adapter(Interface)
 class GateSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The gate booked.")
+    category = _("Booking")
+    description = _("The gate booked.")
 
     def safe_call(self):
         return getattr(self.context, "gate", "")
@@ -33,8 +33,8 @@ class GateSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingDateSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booked date.")
+    category = _("Booking")
+    description = _("The booked date.")
 
     def safe_call(self):
         plone = self.context.restrictedTraverse("@@plone")
@@ -47,8 +47,8 @@ class BookingDateSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingEndDateSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booked end date.")
+    category = _("Booking")
+    description = _("The booked end date.")
 
     def safe_call(self):
         plone = self.context.restrictedTraverse("@@plone")
@@ -61,8 +61,8 @@ class BookingEndDateSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingTimeSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booked time.")
+    category = _("Booking")
+    description = _("The booked time.")
 
     def safe_call(self):
         plone = self.context.restrictedTraverse("@@plone")
@@ -75,8 +75,8 @@ class BookingTimeSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingTimeEndSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking time end.")
+    category = _("Booking")
+    description = _("The booking time end.")
 
     def safe_call(self):
         plone = self.context.restrictedTraverse("@@plone")
@@ -89,8 +89,8 @@ class BookingTimeEndSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingTypeSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking type.")
+    category = _("Booking")
+    description = _("The booking type.")
 
     def safe_call(self):
         return getattr(self.context, "booking_type", "")
@@ -99,8 +99,8 @@ class BookingTypeSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingCodeSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking code.")
+    category = _("Booking")
+    description = _("The booking code.")
 
     def safe_call(self):
         code = self.context.getBookingCode()
@@ -112,8 +112,8 @@ class BookingCodeSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingUrlSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking print url.")
+    category = _("Booking")
+    description = _("The booking print url.")
 
     def safe_call(self):
         return "{folder}/@@prenotazione_print?uid={uid}".format(
@@ -125,8 +125,8 @@ class BookingUrlSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingPrintUrlWithDeleteTokenSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking print url with delete token.")
+    category = _("Booking")
+    description = _("The booking print url with delete token.")
 
     def safe_call(self):
         annotations = IAnnotations(self.context)
@@ -144,8 +144,8 @@ class BookingPrintUrlWithDeleteTokenSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingUserPhoneSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The phone number of the user who made the reservation.")
+    category = _("Booking")
+    description = _("The phone number of the user who made the reservation.")
 
     def safe_call(self):
         return getattr(self.context, "phone", "")
@@ -154,8 +154,8 @@ class BookingUserPhoneSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingUserEmailSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The email address of the user who made the reservation.")
+    category = _("Booking")
+    description = _("The email address of the user who made the reservation.")
 
     def safe_call(self):
         return getattr(self.context, "email", "")
@@ -164,8 +164,8 @@ class BookingUserEmailSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingOfficeContactPhoneSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking office contact phone.")
+    category = _("Booking")
+    description = _("The booking office contact phone.")
 
     def safe_call(self):
         prenotazioni_folder = self.context.getPrenotazioniFolder()
@@ -175,8 +175,8 @@ class BookingOfficeContactPhoneSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingOfficeContactPecSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking office contact pec address.")
+    category = _("Booking")
+    description = _("The booking office contact pec address.")
 
     def safe_call(self):
         prenotazioni_folder = self.context.getPrenotazioniFolder()
@@ -186,8 +186,8 @@ class BookingOfficeContactPecSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingOfficeContactFaxSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking office contact fax.")
+    category = _("Booking")
+    description = _("The booking office contact fax.")
 
     def safe_call(self):
         prenotazioni_folder = self.context.getPrenotazioniFolder()
@@ -197,9 +197,9 @@ class BookingOfficeContactFaxSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingHowToGetToOfficeSubsitution(BaseSubstitution):
 
-    category = _(u"Booking")
+    category = _("Booking")
     description = _(
-        u"The information to reach the office where user book a" " reservation"
+        "The information to reach the office where user book a" " reservation"
     )
 
     def safe_call(self):
@@ -210,10 +210,9 @@ class BookingHowToGetToOfficeSubsitution(BaseSubstitution):
 @adapter(Interface)
 class BookingOfficeCompleteAddressSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
+    category = _("Booking")
     description = _(
-        u"The complete address information of the office where"
-        "user book a reservation"
+        "The complete address information of the office where" "user book a reservation"
     )
 
     def safe_call(self):
@@ -224,8 +223,8 @@ class BookingOfficeCompleteAddressSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingHRDateStartSubstitution(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking human readable date")
+    category = _("Booking")
+    description = _("The booking human readable date")
 
     def safe_call(self):
         # we need something like martedì 8 settembre 2020 alle ore 11:15
@@ -252,8 +251,8 @@ class BookingHRDateStartSubstitution(BaseSubstitution):
 @adapter(Interface)
 class BookingUrlWithDeleteToken(BaseSubstitution):
 
-    category = _(u"Booking")
-    description = _(u"The booking url with delete token")
+    category = _("Booking")
+    description = _("The booking url with delete token")
 
     def safe_call(self):
         annotations = IAnnotations(self.context)